+++ src/actions/TbToggleAction.php
@@ -59,9 +59,6 @@ class TbToggleAction extends CAction
 	/**
 	 * Widgets run function
 	 *
-	 * @param integer $id
-	 * @param string $attribute
-	 *
 	 * @throws CHttpException
 	 */
 	public function run() {
@@ -69,20 +66,19 @@ class TbToggleAction extends CAction
 		$pk = Yii::app()->request->getParam('pk');
 		$attribute = Yii::app()->request->getParam('attribute');
 		
-		if (Yii::app()->getRequest()->isPostRequest) {
-			$model = $this->loadModel($pk);
-			$model->$attribute = ($model->$attribute == $this->noValue) ? $this->yesValue : $this->noValue;
-			$success = $model->save(false, array($attribute));
+		if (!Yii::app()->getRequest()->isPostRequest)
+			throw new CHttpException(400);
 
-			if (Yii::app()->getRequest()->isAjaxRequest) {
-				echo $success ? $this->ajaxResponseOnSuccess : $this->ajaxResponseOnFailed;
-				exit(0);
-			}
-			if ($this->redirectRoute !== null) {
-				$this->getController()->redirect($this->redirectRoute);
-			}
-		} else {
-			throw new CHttpException(Yii::t('zii', 'Invalid request'));
+		$model = $this->loadModel($pk);
+		$model->$attribute = ($model->$attribute == $this->noValue) ? $this->yesValue : $this->noValue;
+		$success = $model->save(false, array($attribute));
+
+		if (Yii::app()->getRequest()->isAjaxRequest) {
+			echo $success ? $this->ajaxResponseOnSuccess : $this->ajaxResponseOnFailed;
+			return;
+		}
+		if ($this->redirectRoute !== null) {
+			$this->getController()->redirect($this->redirectRoute);
 		}
 	}