+++ src/widgets/TbFileUpload.php
@@ -22,6 +22,8 @@ Yii::import('zii.widgets.jui.CJuiInputWidget');
  */
 class TbFileUpload extends CJuiInputWidget
 {
+	private static $_callbackFnPrefix = 'fileupload';
+	
 	/**
 	 * the url to the upload handler
 	 * @var string
@@ -52,10 +54,39 @@ class TbFileUpload extends CJuiInputWidget
 	public $previewImages = true;
 
 	/**
-	 * Whether or not to add the image processing pluing
+	 * Whether or not to add the image processing plugin
 	 */
 	public $imageProcessing = true;
 
+	/**
+	 * Stores callback functions JS code, reffering do jQuery-File-Upload documentation
+	 * Use these values for array keys:
+	 * 	- add, 
+	 *  - submit, 
+	 *  - send, 
+	 *  - done, 
+	 *  - fail, 
+	 *  - always, 
+	 *  - progress, 
+	 *  - progressall, 
+	 *  - start, 
+	 *  - stop
+	 * 	- change, 
+	 *  - pase, 
+	 *  - drop, 
+	 *  - dragover, 
+	 *  - chunksend, 
+	 *  - chunkdone, 
+	 *  - chunkfail, 
+	 *  - chunkalways
+	 *  Another part of callback function name will be added from code.
+	 * 
+	 * See link below for more details:
+	 * @link https://github.com/blueimp/jQuery-File-Upload/wiki/Options#callback-options Callback options
+	 * @var string[] 
+	 */
+	public $callbacks = array( );
+        
 	/**
 	 * @var string name of the form view to be rendered
 	 */
@@ -102,7 +133,7 @@ class TbFileUpload extends CJuiInputWidget
 	public function run()
 	{
 
-		list($name, $id) = $this->resolveNameID();
+		list($name) = $this->resolveNameID();
 
 		$this->htmlOptions['id'] = $this->id.'-'.($this->hasModel() ? get_class($this->model) : 'fileupload') . '-form';
 
@@ -184,7 +215,7 @@ class TbFileUpload extends CJuiInputWidget
         $booster->registerAssetJs('fileupload/jquery.fileupload-ui.js');
 
 		$options = CJavaScript::encode($this->options);
-		Yii::app()->clientScript->registerScript(__CLASS__ . '#' . $id, "jQuery('#{$id}').fileupload({$options});");
+		Yii::app()->clientScript->registerScript(__CLASS__ . '#' . $id, "jQuery('#{$id}').fileupload({$options}){$this->generateCallbackBindJSString()};");
 	}
 
 	/**
@@ -192,7 +223,7 @@ class TbFileUpload extends CJuiInputWidget
 	 *
 	 * @param CModel $model
 	 * @param string $attribute
-	 * @param null $property
+	 * @param string $property
 	 *
 	 * @return string property's value or null
 	 */
@@ -208,5 +239,17 @@ class TbFileUpload extends CJuiInputWidget
 			}
 		}
 		return isset($ret) ? $ret : null;
+	}
+	
+	private function generateCallbackBindJSString() {
+		if ( count($this->callbacks) > 0 ) {
+			$output = '';
+			foreach ( $this->callbacks as $callbackName => $fnCode ) {
+				$output .= '.bind("' . self::$_callbackFnPrefix
+					. $callbackName . '", ' . $fnCode . ')';
+                     }
+			return $output;
+		}
+		return '';
 	}
 }