+++ src/widgets/TbWidget.php
@@ -48,19 +48,23 @@ abstract class TbWidget extends CWidget {
 	}
 
 	/**
-	 * 
-	 * @param string $context
+	 *
+	 * @param bool|string $context
+	 *
+	 * @return bool
 	 */
-	protected function isValidContext($cotext = false) {
-		if($cotext)
+	protected function isValidContext($context = false) {
+		if($context)
 			return defined(get_called_class().'::CTX_'.strtoupper($context));
 		else
 			return defined(get_called_class().'::CTX_'.strtoupper($this->context));
 	}
-	
+
 	/**
-	 * 
-	 * @param string $context
+	 *
+	 * @param bool|string $context
+	 *
+	 * @return mixed
 	 */
 	protected function getContextClass($context = false) {
 		if($context)